requirements.txt
# RAG Brain - Local Multimodal RAG System
# CPU-optimized requirements

# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# ML/AI - Local CPU models
sentence-transformers==3.3.1
transformers==4.46.2
torch==2.5.1+cpu
torchvision==0.20.1+cpu
accelerate==1.1.1
Pillow==11.0.0

# Vector Store & Database
faiss-cpu==1.8.0
sqlalchemy==2.0.35
aiosqlite==0.20.0

# MLflow
mlflow==2.18.0

# File Processing
python-docx==1.1.2
PyPDF2==3.0.1
chardet==5.2.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
tenacity==9.0.0

# Development/Testing
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
